+++ providers/jdbc_connection_pool.rb
@@ -41,3 +41,18 @@ action :create do
     code asadmin_command(command.join(' '))
   end
 end
+
+action :delete do
+  command = []
+  command << "delete-jdbc-connection-pool"
+  command << "--target" << new_resource.target if new_resource.target
+  command << "--cascade=true"
+  command << new_resource.name
+
+  bash "asadmin_delete_jdbc_connection_pool #{new_resource.name}" do
+    only_if "#{asadmin_command('list-jdbc-connection-pools')} | grep -x -- '#{new_resource.name}'"
+    user node['glassfish']['user']
+    group node['glassfish']['group']
+    code asadmin_command(command.join(' '))
+  end
+end